Add download service to DownloadFileService class and fix issues

+++ CrossPlatformDownloadManager.Utils/Constants.cs
@@ -5,10 +5,10 @@ namespace CrossPlatformDownloadManager.Utils;
 public static class Constants
 {
     // File Size
-    public const double KB = 1024;
-    public const double MB = KB * 1024;
-    public const double GB = MB * 1024;
-    public const double TB = GB * 1024;
+    public const long KB = 1024;
+    public const long MB = KB * 1024;
+    public const long GB = MB * 1024;
+    public const long TB = GB * 1024;
 
     // Turn off computer modes
     public static readonly List<string> TurnOffComputerModes = Enum.GetNames(typeof(TurnOffComputerMode))
